Do not show truncated logs in log output
This was supposed to be a usability feature: if the error from
a task is too long, the last N lines were displayed and Game
pointed user to the full log above. In practice it turned out
that if the task's output is long, then the probability of
issue being printed in the last N lines is low.

So if the log is too long, don't print it at all.